Create finalyearproject
“CLUSTERING OF WIRELESS SENSOR NETWORKS USING PAM ALGORITHM”
Balanced Centroid is that the centroid within the cluster on a central location on the basis of density of the nodes irrespective to the cluster area center point. The N numbers of nodes are initially evaluated for the cluster formation. The distance between the nodes is calculated using Euclidean distance. The nodes are divided into K number of clusters. For every cluster K, its N number of nodes are evaluated using distance formula again to search out the centroid by initializing the first stage of centroid election with random co-ordinates. Then the balanced CH (cluster head) is elected by the algorithm for a cluster K with N number of nodes. The proposed solution is predicated on the cluster algorithm K-medoids, during which centroid is chosen on the basis of medoids technique of statistics. K-medoids is an algorithm uses portioning around medoids which in real select centroids round the actual center of the clusters which is given by X-center (Xc) and Y-center (Yc) respectively. The proposed algorithm for the WSN clustering has performed well with the WSN simulations in NS2. Euclidean distance is employed to compute the gap between the 2 nodes under this simulation. The cluster on which the results were tested contained 15 no. of nodes and this number can vary accordingly.